+++ R/fitEllipse.R
@@ -1,35 +1,36 @@
 #' Fit a multivariate normal distribution to x and y data using jags
 #' 
-#' This function contains and defines the jags model script used to fit a
-#' bivariate normal distribution to a vector of x and y data. Although not
-#' intended for direct calling by users, it presents a quick way to fit a model
+#' This function contains and defines the jags model script used to fit a 
+#' bivariate normal distribution to a vector of x and y data. Although not 
+#' intended for direct calling by users, it presents a quick way to fit a model 
 #' to a single group of data. Advanced users should be able to manipulate the 
 #' contained jags model to fit more complex models using different likelihoods, 
-#' such as multivariate lognormal distributions, multivariate gamma
+#' such as multivariate lognormal distributions, multivariate gamma 
 #' distributions etc...
 #' 
 #' @param x a vector of data representing the x-axis
 #' @param y a vector of data representing the y-axis
-#' @param parms a list containing four items providing details of the
-#'  \code{\link[rjags]{rjags}} run to be sampled.
-#' \itemize{
-#'    \item {n.iter}{The number of iterations to sample}
-#'    \item {n.burnin}{The number of iterations to discard as a burnin from the
-#'                      start of sampling.}
-#'    \item {n.thin}{The number of samples to thin by.}
-#'    \item {n.chains}{The number of chains to fit.}
-#' }
-#' @param priors a list of three items specifying the priors to be passed to 
-#' the jags model.
-#' \itemize{
-#'    \item {R}{The scaling vector for the diagonal of Inverse Wishart
-#'    distribution prior on the covariance matrix Sigma. Typically 
-#'    set to a 2x2 matrix matrix(c(1, 0, 0, 1), 2, 2).}
-#'    \item {k}{The degrees of freedom of the Inverse Wishart distribution for 
-#'    the covariance matrix Sigma. Typically set to the dimensionality of Sigma,
-#'    which in this bivariate case is 2.}
-#'    \item {tau}{The precision on the normal prior on the means mu.}
-#' }
+#' @param parms a list containing four items providing details of the 
+#'   \code{\link[rjags]{rjags}} run to be sampled. \itemize{ \item {n.iter}{The 
+#'   number of iterations to sample} \item {n.burnin}{The number of iterations 
+#'   to discard as a burnin from the start of sampling.} \item {n.thin}{The 
+#'   number of samples to thin by.} \item {n.chains}{The number of chains to 
+#'   fit.} }
+#' @param priors a list of three items specifying the priors to be passed to the
+#'   jags model. \itemize{ \item {R}{The scaling vector for the diagonal of 
+#'   Inverse Wishart distribution prior on the covariance matrix Sigma. 
+#'   Typically set to a 2x2 matrix matrix(c(1, 0, 0, 1), 2, 2).} \item {k}{The 
+#'   degrees of freedom of the Inverse Wishart distribution for the covariance 
+#'   matrix Sigma. Typically set to the dimensionality of Sigma, which in this 
+#'   bivariate case is 2.} \item {tau}{The precision on the normal prior on the 
+#'   means mu.} }
+#' @param id a character string to prepend to the raw saved jags model output. 
+#'   This is typically passed on from the calling function 
+#'   \code{\link{siberMVN}} and identifies the community and group with an 
+#'   integer labelling system. Defaults to NULL which will prevent the output 
+#'   object being saved even if \code{parms$save.output} is set to \code{TRUE}.
+#'   The file itself will be saved to the user-specified location via
+#'   \code{parms$save.dir}.
 #' @return A mcmc.list object of posterior samples created by jags.
 #' @examples
 #' x <- stats::rnorm(50)
@@ -47,7 +48,7 @@
 #' 
 #' @export
 
-fitEllipse <- function (x, y, parms, priors) 
+fitEllipse <- function (x, y, parms, priors, id = NULL) 
 {
 
   # ----------------------------------------------------------------------------
@@ -110,13 +111,16 @@ fitEllipse <- function (x, y, parms, priors)
   parameters <- c("mu","Sigma2")
   
   model <- rjags::jags.model(textConnection(modelstring),
-                             data = jags.data, n.chains = 2)
+                             data = jags.data, n.chains = parms$n.chains)
   
   output <- rjags::coda.samples(model = model,
                                 variable.names = c("mu",'Sigma2'),
                                 n.iter = parms$n.iter,
-                                thin = 10
+                                thin = parms$n.thin
                                 )
+  
+  if (!is.null(id) & parms$save.output) {
+    save(output, file = paste0(parms$save.dir, "/jags_output_", id, ".RData"))}
   
   #print(summary(output))
   #print(dim(output[[1]]))